+++ backend/src/main/java/net/es/oscars/pss/svc/PSSAdapter.java
@@ -6,6 +6,7 @@ import net.es.oscars.app.exc.NotReadyException;
 import net.es.oscars.app.exc.NsiException;
 import net.es.oscars.app.exc.PSSException;
 import net.es.oscars.app.props.PssProperties;
+import net.es.oscars.app.syslog.Syslogger;
 import net.es.oscars.dto.pss.cmd.*;
 import net.es.oscars.dto.pss.st.ConfigStatus;
 import net.es.oscars.dto.pss.st.LifecycleStatus;
@@ -35,7 +36,6 @@ import java.util.Optional;
 import java.util.concurrent.*;
 import java.util.stream.Collectors;
 
-
 @Component
 @Slf4j
 public class PSSAdapter {
@@ -47,12 +47,12 @@ public class PSSAdapter {
     private LogService logService;
     private ConnService connService;
     private PSSQueuer queuer;
-
     private TopoService topoService;
+    private Syslogger syslogger;
 
     @Autowired
     public PSSAdapter(PSSProxy pssProxy, RouterCommandsRepository rcr, CommandHistoryRepository historyRepo,
-                      NsiService nsiService, ConnService connService, PSSQueuer queuer,
+                      NsiService nsiService, ConnService connService, PSSQueuer queuer, Syslogger syslogger,
                       TopoService topoService, LogService logService, PssProperties properties) {
         this.pssProxy = pssProxy;
         this.rcr = rcr;
@@ -63,6 +63,7 @@ public class PSSAdapter {
         this.nsiService = nsiService;
         this.logService = logService;
         this.properties = properties;
+        this.syslogger = syslogger;
     }
 
     public State processTask(Connection conn, CommandType commandType, State intent) {
@@ -97,17 +98,22 @@ public class PSSAdapter {
 
     public State build(Connection conn) throws PSSException, NotReadyException {
         log.info("building " + conn.getConnectionId());
+        syslogger.sendSyslog( "OSCARS BUILD STARTED : " + conn.getConnectionId());
+
         List<Command> commands = this.configCommands(conn, CommandType.BUILD);
         List<CommandStatus> stable = this.getStableStatuses(commands);
         Instant now = Instant.now();
 
         State result = State.ACTIVE;
         for (CommandStatus st : stable) {
+            RouterCommands rc = this.existing(conn.getConnectionId(), st.getDevice(), CommandType.BUILD);
+
             RouterCommandHistory rch = RouterCommandHistory.builder()
                     .connectionId(conn.getConnectionId())
                     .date(now)
                     .deviceUrn(st.getDevice())
                     .commands(st.getCommands())
+                    .templateVersion(rc.getTemplateVersion())
                     .output(st.getOutput())
                     .configStatus(st.getConfigStatus())
                     .type(CommandType.BUILD)
@@ -124,6 +130,9 @@ public class PSSAdapter {
                         .username("system")
                         .build();
                 logService.logEvent(conn.getConnectionId(), ev);
+
+                // Send Syslog Message
+                syslogger.sendSyslog( "OSCARS BUILD FAILED : " + conn.getConnectionId());
             } else {
                 Event ev = Event.builder()
                         .connectionId(conn.getConnectionId())
@@ -133,6 +142,9 @@ public class PSSAdapter {
                         .username("system")
                         .build();
                 logService.logEvent(conn.getConnectionId(), ev);
+
+                // Send Syslog Message
+                syslogger.sendSyslog( "OSCARS BUILD ENDED SUCCESSFULLY : " + conn.getConnectionId());
             }
 
 
@@ -143,20 +155,26 @@ public class PSSAdapter {
 
     public State dismantle(Connection conn) throws PSSException, NotReadyException {
         log.info("dismantling " + conn.getConnectionId());
+        syslogger.sendSyslog( "OSCARS DISMANTLE STARTED : " + conn.getConnectionId());
+
         List<Command> commands = this.configCommands(conn, CommandType.DISMANTLE);
         List<CommandStatus> stable = this.getStableStatuses(commands);
         Instant now = Instant.now();
         State result = State.WAITING;
         for (CommandStatus st : stable) {
+            RouterCommands rc = this.existing(conn.getConnectionId(), st.getDevice(), CommandType.DISMANTLE);
+
             RouterCommandHistory rch = RouterCommandHistory.builder()
                     .connectionId(conn.getConnectionId())
                     .date(now)
                     .deviceUrn(st.getDevice())
                     .commands(st.getCommands())
+                    .templateVersion(rc.getTemplateVersion())
                     .output(st.getOutput())
                     .configStatus(st.getConfigStatus())
                     .type(CommandType.DISMANTLE)
                     .build();
+
             historyRepo.save(rch);
             if (st.getConfigStatus().equals(ConfigStatus.ERROR)) {
                 result = State.FAILED;
@@ -168,6 +186,7 @@ public class PSSAdapter {
                         .username("system")
                         .build();
                 logService.logEvent(conn.getConnectionId(), ev);
+                syslogger.sendSyslog( "OSCARS DISMANTLE FAILED : " + conn.getConnectionId());
             } else {
                 Event ev = Event.builder()
                         .connectionId(conn.getConnectionId())
@@ -177,6 +196,9 @@ public class PSSAdapter {
                         .username("system")
                         .build();
                 logService.logEvent(conn.getConnectionId(), ev);
+
+                // Send Syslog Message
+                syslogger.sendSyslog( "OSCARS DISMANTLE ENDED SUCCESSFULLY : " + conn.getConnectionId());
             }
         }
         this.triggerNsi(conn, result);
@@ -195,12 +217,6 @@ public class PSSAdapter {
 
     }
 
-
-
-
-
-
-
     public List<CommandStatus> getStableStatusesSerial(List<Command> commands) throws PSSException {
         List<CommandResponse> responses = serialSubmit(commands);
         List<String> commandIds = responses.stream()
@@ -208,7 +224,6 @@ public class PSSAdapter {
                 .collect(Collectors.toList());
         return pollUntilStable(commandIds);
     }
-
 
     public List<CommandStatus> getStableStatuses(List<Command> commands) throws PSSException {
         try {
@@ -222,8 +237,6 @@ public class PSSAdapter {
             throw new PSSException("interrupted");
         }
     }
-
-
 
     public List<CommandStatus> pollUntilStable(List<String> commandIds)
             throws PSSException {
@@ -310,7 +323,6 @@ public class PSSAdapter {
         return responses;
     }
 
-
     public List<CommandStatus> pollStatuses(List<String> commandIds)
             throws InterruptedException, ExecutionException {
         List<CommandStatus> statuses = new ArrayList<>();
@@ -334,8 +346,7 @@ public class PSSAdapter {
         return statuses;
     }
 
-
-    public List<Command> configCommands(Connection conn, CommandType ct) throws PSSException, NotReadyException {
+   public List<Command> configCommands(Connection conn, CommandType ct) throws PSSException, NotReadyException {
         log.info("gathering "+ct+" commands for " + conn.getConnectionId());
         List<Command> commands = new ArrayList<>();
 
@@ -362,7 +373,6 @@ public class PSSAdapter {
 
         return commands;
     }
-
 
     public RouterCommands existing(String connId, String deviceUrn, CommandType commandType) {
         List<RouterCommands> existing = rcr.findByConnectionIdAndDeviceUrn(connId, deviceUrn);