+++ pss/src/main/java/net/es/oscars/pss/app/StartupComponent.java
@@ -1,0 +1,7 @@
+package net.es.oscars.pss.app;
+
+import net.es.oscars.pss.app.exc.StartupException;
+
+public interface StartupComponent {
+    void startup() throws StartupException;
+}