+++ pss/src/main/java/net/es/oscars/pss/app/exc/StartupException.java
@@ -1,0 +1,7 @@
+package net.es.oscars.pss.app.exc;
+
+public class StartupException extends Exception {
+    public StartupException(String msg) {
+        super(msg);
+    }
+}