+++ pss/src/main/java/net/es/oscars/pss/app/syslog/Syslogger.java
@@ -1,0 +1,12 @@
+package net.es.oscars.pss.app.syslog;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+public class Syslogger {
+    public void sendSyslog(String message)  {
+        log.info(message);
+    }
+}